requirements.txt
REQUIREMENTS


fastapi
uvicorn
google-genai
pydantic